+++ code/client/src/models/admin/ArticleVO.js
@@ -1,6 +1,0 @@
-class ArticleVO {
-  constructor() {
-      
-  }
-}
-export default ArticleVO