+++ code/client/src/pages/admin/article/ArticleManage.jsx
@@ -1,33 +1,25 @@
-import React, { Component } from "react";
-import Manage from "../../../components/Manage/Manage";
-import { message as Message } from "antd";
-import ArticleApi from "../../../api/article/ArticleApi";
+import React, { Component } from 'react'
+import Manage from '@/components/Manage/Manage'
+import ArticleService from '@/service/admin/article/ArticleService'
 
 class ArticleManage extends Component {
-    state = {
-        tableList: [],
-    };
-    componentDidMount() {
-        //获取所有 Article 信息
-        new ArticleApi().getAll().then((resolve, reject) => {
-            let { status, message, data } = resolve.data;
-            if (status === 100) {
-                this.setState({
-                    tableList: data,
-                });
-            } else {
-                Message.warning(message);
-            }
-        });
-    }
-    render() {
-        let { tableList } = this.state;
-        return tableList.length !== 0 ? (
-            <Manage tableList={this.state.tableList}></Manage>
-        ) : (
-            '后台数据失联啦，请等待。。。'
-        );
-    }
+  state = {
+    tableList: [],
+  }
+  componentDidMount() {
+    let articleService = new ArticleService()
+    articleService.getArticleTableList().then((data) => {
+      this.setState({ tableList: data })
+    })
+  }
+  render() {
+    let { tableList } = this.state
+    return tableList.length !== 0 ? (
+      <Manage tableList={tableList}></Manage>
+    ) : (
+      '后台数据失联啦，请等待。。。'
+    )
+  }
 }
 
-export default ArticleManage;
+export default ArticleManage